Day 21st of April ( idk, idk2 )
Added border PacMan shenanigans, added some SFXs,  added score, made the game losable, made the player take dmg when they collide with an enemy, made the thingies on the mid's sprite change when they take damage, added a speed boost for when the player repairs a turret and lastly made a thing that let the player choose the turret they want to place. A lot of stuff ngl. Was able to achieve all the goals for today :)